pyproject.toml
[project]
name = "autoformation"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "argon2-cffi>=25.1.0",
    "fastapi-mail>=1.6.1",
    "fastapi[standard]>=0.128.0",
    "passlib[bcrypt]>=1.7.4",
    "pg8000>=1.31.5",
    "psycopg2-binary>=2.9.11",
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.2.1",
    "python-jose[cryptography]>=3.5.0",
    "sqlalchemy>=2.0.46",
]
